fw_update: check existing candidate image before download
Update fw_update thread to check whether the requested firmware update
image is already present in candidate slot. If so, download is skipped and
we attempt to reboot into existing candidate slot image.

Note that the call to fw_update_validate is now removed from the post
download flow. fw_update_validate was previously a no-op for most platform
ports, and for those that did implement checks (esp_idf), the functionality
was moved to fw_update_post_download, which is still called following an
image download.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>